model_training/LLM_Model_Finetuning/requirements.txt
pandas
requests
python-dotenv
numpy
scikit-learn
datasets
torch
transformers
accelerate>=0.26.0
peft
bitsandbytes
tabulate
tiktoken
protobuf
sentencepiece
